Correcciones toma de muestra Correciones NumberPage Decimales

+++ src/ni/org/ics/estudios/appmovil/cohortefamilia/activities/enterdata/NuevaEncuestaDatosPartoBBActivity.java
@@ -285,7 +285,7 @@ public class NuevaEncuestaDatosPartoBBActivity  extends FragmentActivity impleme
             if (!page.getData().isEmpty() && clase.equals("class ni.org.ics.estudios.appmovil.wizard.model.NumberPage")) {
                 NumberPage np = (NumberPage) page;
                 String valor = np.getData().getString(NumberPage.SIMPLE_DATA_KEY);
-                if((np.ismValRange() && (np.getmGreaterOrEqualsThan() > Integer.valueOf(valor) || np.getmLowerOrEqualsThan() < Integer.valueOf(valor)))
+                if((np.ismValRange() && (np.getmGreaterOrEqualsThan() > Double.valueOf(valor) || np.getmLowerOrEqualsThan() < Double.valueOf(valor)))
                         || (np.ismValPattern() && !valor.matches(np.getmPattern()))){
                     cutOffPage = i;
                     break;
